backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
Pillow==10.0.1
python-decouple==3.8
django-filter==23.3
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
